+++ Dockerfile
@@ -9,6 +9,7 @@ COPY go.mod .
 COPY go.sum .
 COPY main.go .
 COPY pkg ./pkg
+COPY internal ./internal
 
 RUN go mod download
