+++ internal/driver/controllerserver.go
@@ -1,4 +1,4 @@
-package linodebs
+package driver
 
 import (
 	"context"
