+++ internal/driver/identityserver.go
@@ -12,7 +12,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package linodebs
+package driver
 
 import (
 	csi "github.com/container-storage-interface/spec/lib/go/csi"
