+++ internal/driver/nodeserver_all.go
@@ -1,6 +1,6 @@
 //go:build !windows
 
-package linodebs
+package driver
 
 import (
 	"context"
